fix: prompt, report 테이블 구조 변경
